Big refactors for PyPi
