+++ pyproject.toml
@@ -1,0 +1,26 @@
+[project]
+name = "pious"
+version = "0.0.0.dev1"
+authors = [
+    { name="Ben Kushigian", email="bkushigian@gmail.com" }
+]
+description = "The PioSOLVER Utility Suite"
+readme = "README.md"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+dependencies = [
+    "matplotlib>=3.7.1",
+    "pandas>=2.0.3",
+    "mplcursors>=0.5.1"
+]
+
+[project.urls]
+Issues = "https://github.com/bkushigian/Pious/issues"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
