+++ src/pious/util/rebuild_utils.py
@@ -6,12 +6,13 @@ This is to replace the broken functionality of `solve_all_spots` from UPI
 """
 
 from argparse import ArgumentParser
-from pyosolver import PYOSolver, Node
-from pio_utils import Line, filter_lines, is_flop, is_turn, is_nonterminal
-from progress_bar import progress_bar
-from pio_utils import make_solver, FLOP, TURN
 from os import path as osp
 import time
+
+from pious.pyosolver import PYOSolver, Node
+from pious.pio_utils import Line, filter_lines, is_flop, is_turn, is_nonterminal
+from pious.pio_utils import make_solver, FLOP, TURN
+from pious.util.progress_bar import progress_bar
 
 
 def rebuild_and_resolve(solver: PYOSolver, lock_turns=True, lines=None, accuracy=0.05):
